Logger queues are now singletons
